% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordionItem}
\alias{accordionItem}
\title{AdminLTE2 accordion item}
\usage{
accordionItem(..., id, title = NULL, color = NULL)
}
\arguments{
\item{...}{text to write in the item.}

\item{id}{unique item id.}

\item{title}{item title.}

\item{color}{item color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create an accordion item to put inside an accordion container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
