% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{attachmentBlock}
\alias{attachmentBlock}
\title{AdminLTE2 attachment container}
\usage{
attachmentBlock(..., src = NULL, title = NULL, titleUrl = NULL)
}
\arguments{
\item{...}{any element.}

\item{src}{url or path to the image.}

\item{title}{attachment title.}

\item{titleUrl}{external link.}
}
\description{
Create an attachment container, nice to wrap articles...
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Attachment example",
      attachmentBlock(
       src = "https://adminlte.io/themes/AdminLTE/dist/img/photo1.png",
       title = "Test",
       titleUrl = "https://google.com",
       "This is the content"
      )
     )
    ),
    title = "AttachmentBlock"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
