% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxComment}
\alias{boxComment}
\title{AdminLTE2 box comment}
\usage{
boxComment(..., image, title, date = NULL)
}
\arguments{
\item{...}{comment content.}

\item{image}{author image, if any.}

\item{title}{comment title.}

\item{date}{date of publication.}
}
\description{
Create box comment
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
