% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxDropdown}
\alias{boxDropdown}
\title{Create a box dropdown item list}
\usage{
boxDropdown(..., icon = shiny::icon("wrench"))
}
\arguments{
\item{...}{Slot for \link{boxDropdownItem}.}

\item{icon}{Dropdown menu icon. Expect \code{\link[shiny]{icon}}.}
}
\description{
Can be used to add dropdown items to a boxtool.
}
