% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxDropdownItem}
\alias{boxDropdownItem}
\title{Create a box dropdown item}
\usage{
boxDropdownItem(url, name)
}
\arguments{
\item{url}{Target url or page.}

\item{name}{Menu name.}
}
\description{
Create a box dropdown item
}
