% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxPlusLabel}
\alias{boxPlusLabel}
\title{Create a label for \link{boxPlus}}
\usage{
boxPlusLabel(text, status, style = "default")
}
\arguments{
\item{text}{Label text. In practice only few letters or a number.}

\item{status}{label color status. See \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}

\item{style}{label border style: "default" (rounded angles), "circle" or "square".}
}
\description{
Create a label for \link{boxPlus}
}
\seealso{
\link{dashboardLabel}
}
