% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxProfile}
\alias{boxProfile}
\title{AdminLTE2 box profile}
\usage{
boxProfile(..., src = NULL, title, subtitle = NULL, bordered = FALSE)
}
\arguments{
\item{...}{any element such as \link{boxProfileItem}.}

\item{src}{profile image, if any.}

\item{title}{title.}

\item{subtitle}{subtitle.}

\item{bordered}{Whether the container should have a border or not. FALSE by default.}
}
\description{
Create box profile
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Box with profile",
      status = "primary",
      boxProfile(
       src = "https://adminlte.io/themes/AdminLTE/dist/img/user4-128x128.jpg",
       title = "Nina Mcintire",
       subtitle = "Software Engineer",
       bordered = TRUE,
       boxProfileItem(
        title = "Followers",
        description = 1322
       ),
       boxProfileItem(
        title = "Following",
        description = 543
       ),
       boxProfileItem(
        title = "Friends",
        description = 13287
       )
      )
     )
    ),
    title = "boxProfile"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
