% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxProfileItem}
\alias{boxProfileItem}
\title{AdminLTE2 box profile item}
\usage{
boxProfileItem(title = NULL, description = NULL)
}
\arguments{
\item{title}{item title.}

\item{description}{item info.}
}
\description{
Create box profile item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
