% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxProfileItemList}
\alias{boxProfileItemList}
\title{AdminLTE2 box profile item container}
\usage{
boxProfileItemList(..., bordered = FALSE)
}
\arguments{
\item{...}{slot for boxProfileItem.}

\item{bordered}{Whether the container should have a border or not. FALSE by default.}
}
\description{
Create box profile item list
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
