% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxSidebar}
\alias{boxSidebar}
\title{Create a sidebar for a box}
\usage{
boxSidebar(
  ...,
  id = NULL,
  width = "25\%",
  background = "#333a40",
  startOpen = FALSE,
  icon = shiny::icon("cogs")
)
}
\arguments{
\item{...}{Sidebar content.}

\item{id}{Unique sidebar id. Useful if you want to use \link{updateBoxSidebar}.}

\item{width}{Sidebar width in percentage. 25\% by default. A character value of any width CSS understands (e.g. "100px").}

\item{background}{Sidebar background color. Dark by default.}

\item{startOpen}{Whether the sidebar is open at start. FALSE by default.}

\item{icon}{Sidebar icon. Expect \code{\link[shiny]{icon}}.}
}
\description{
To insert in the sidebar slot of \link{box}.
}
