% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{carouselItem}
\alias{carouselItem}
\title{AdminLTE2 carousel item}
\usage{
carouselItem(..., caption = "")
}
\arguments{
\item{...}{Element such as images, iframe, ...}

\item{caption}{Item caption.}
}
\description{
This creates a carousel item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
