% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{dashboardBadge}
\alias{dashboardBadge}
\title{AdminLTE2 badge}
\usage{
dashboardBadge(..., status = "primary")
}
\arguments{
\item{...}{Any html text element.}

\item{status}{label status.}
}
\description{
Create a badge. Contrary to the label
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     dashboardBadge("Label 1", status = "info"),
     actionButton(
      inputId = "badge", 
      label = "Hello", 
      icon = NULL, 
      width = NULL, 
      dashboardBadge(1, status = "orange")
     )
    )
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
