% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{dashboardFooter}
\alias{dashboardFooter}
\title{Dashboard Footer}
\usage{
dashboardFooter(left = NULL, right = NULL)
}
\arguments{
\item{left}{Left text.}

\item{right}{Right text.}
}
\description{
This creates a dashboard footer
}
\examples{
if (interactive()) {
library(shiny)
library(shinydashboard)
library(shinydashboardPlus)

shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(),
    sidebar = dashboardSidebar(),
    body = dashboardBody(),
    footer = dashboardFooter(
     left = "By Divad Nojnarg",
     right = "Zurich, 2019"
    ),
    title = "DashboardPage"
  ),
  server = function(input, output) { }
)
}
}
