% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardLabel.R
\name{dashboardLabel}
\alias{dashboardLabel}
\title{AdminLTE2 label}
\usage{
dashboardLabel(..., status = "primary", style = "default")
}
\arguments{
\item{...}{any text.}

\item{status}{label status: "danger", "success", "info", "primary", "warning".}

\item{style}{label border style: "default" (rounded angles), "circle" or "square".}
}
\description{
Create a label
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     dashboardLabel("Label 1", status = "info"),
     dashboardLabel("Label 2", status = "danger", style = "circle"),
     dashboardLabel("Label 3", status = "success", style = "square")
    )
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
