% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardPagePlus.R
\name{dashboardPagePlus}
\alias{dashboardPagePlus}
\title{Dashboard Page with a right sidebar}
\usage{
dashboardPagePlus(
  header,
  sidebar,
  body,
  controlbar = NULL,
  footer = NULL,
  title = NULL,
  skin = c("blue", "blue-light", "black", "black-light", "purple", "purple-light",
    "green", "green-light", "red", "red-light", "yellow", "yellow-light", "midnight"),
  preloader = FALSE,
  duration = 2,
  md = FALSE,
  options = NULL
)
}
\arguments{
\item{header}{A header created by \code{dashboardHeaderPlus}.}

\item{sidebar}{A sidebar created by \code{\link[shinydashboard]{dashboardSidebar}} or 
\code{\link{dashboardSidebarPlus}}.}

\item{body}{A body created by \code{\link[shinydashboard]{dashboardBody}}.}

\item{controlbar}{A right sidebar created by \link{dashboardControlbar}. NULL by
default.}

\item{footer}{A footer created by \link{dashboardFooter}.}

\item{title}{A title to display in the browser's title bar. If no value is
provided, it will try to extract the title from the
\code{dashboardHeaderPlus}.}

\item{skin}{A color theme. See \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.
If the skin is light, the sidebar will have a light background.}

\item{preloader}{Whether to enable a page loader. FALSE by default.}

\item{duration}{Loader duration in seconds. 2s by default.}

\item{md}{Whether to enable material design. Experimental...}

\item{options}{Extra option to overwrite the vanilla AdminLTE configuration. See 
\url{https://adminlte.io/themes/AdminLTE/documentation/index.html#adminlte-options}.
Expect a list.}
}
\description{
This creates a dashboard page for use in a Shiny app.
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
   ui = dashboardPagePlus(
     collapse_sidebar = TRUE,
     options = list(sidebarExpandOnHover = TRUE),
     header = dashboardHeaderPlus(
      enable_rightsidebar = TRUE,
      rightSidebarIcon = "gears"
     ),
     sidebar = dashboardSidebar(),
     body = dashboardBody(),
     rightsidebar = rightSidebar(),
     title = "DashboardPage"
   ),
   server = function(input, output) { }
 )
}
}
\seealso{
\code{\link{dashboardHeaderPlus}}, \code{\link[shinydashboard]{dashboardSidebar}},
  \code{\link[shinydashboard]{dashboardBody}}.
}
