% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeaderPlus.R
\name{dashboardUserItem}
\alias{dashboardUserItem}
\title{Create a dashboard user profile item}
\usage{
dashboardUserItem(item, width)
}
\arguments{
\item{item}{HTML Tag.}

\item{width}{Item width between 1 and 12.}
}
\description{
This can be inserted in a \code{\link{dashboardUser}}.
}
