% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeaderPlus.R
\name{dropdownBlock}
\alias{dropdownBlock}
\title{Create a dropdown block to place in a dashboard header}
\usage{
dropdownBlock(..., id, icon = NULL, title = NULL,
  badgeStatus = "danger")
}
\arguments{
\item{...}{Items to put in the menu.}

\item{id}{Dropdown block id.}

\item{icon}{An icon to display in the header. By default, the icon is
automatically selected depending on \code{type}, but it can be overriden
with this argument.}

\item{title}{Dropdown block title.}

\item{badgeStatus}{Dropdown badge status.}
}
\description{
Create a dropdown block to place in a dashboard header
}
\seealso{
\code{\link{dashboardHeaderPlus}} for example usage.
}
