% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced-boxes.R
\name{dropdownDivider}
\alias{dropdownDivider}
\title{Create a box dropdown divider}
\usage{
dropdownDivider()
}
\description{
Create a box dropdown divider
}
\note{
Useful to separate 2 sections of dropdown items.
}
