% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced-boxes.R
\name{dropdownItem}
\alias{dropdownItem}
\title{Create a box dropdown item}
\usage{
dropdownItem(target = NULL, name = NULL)
}
\arguments{
\item{target}{Target url or page.}

\item{name}{Menu name.}
}
\description{
Create a box dropdown item
}
