% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{dropdownItemList}
\alias{dropdownItemList}
\title{Create a box dropdown item list}
\usage{
dropdownItemList(..., icon = shiny::icon("wrench"))
}
\arguments{
\item{...}{Slot for dropdownItem.}

\item{icon}{Dropdown menu icon.}
}
\description{
Can be used to add dropdown items to a boxtool.
}
