% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_box.R
\name{flip_box}
\alias{flip_box}
\title{flip_box}
\usage{
flip_box(
  id,
  front_content,
  back_content,
  front_text = "More Info",
  back_text = "Back to Front"
)
}
\arguments{
\item{id}{the flipbox id}

\item{front_content}{ui for the front of the flip box}

\item{back_content}{ui for the back of the flip box}

\item{front_text}{text for front button}

\item{back_text}{text for back button}
}
\value{
HTML tagList
}
\description{
AdminLTE flipBox.
}
\section{Warning}{
 this function is experimental.
}

\examples{
library(shiny)
library(shinydashboard)
library(shinydashboardPlus)
shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
      flip_box(
        id = "id",
        front_content = div(
          class = "text-center",
          img(
            src = "https://image.flaticon.com/icons/svg/149/149076.svg",
            height = "300px",
            width = "100\%"
          )
        ),
        back_content = div(
          class = "text-center",
          height = "300px",
          width = "100\%",
          h1("Details...."),
          p("More information....")
        )
      )
    )
  ),
  
  server = function(input, output, session) {
  }
)
   
}
\references{
\url{https://rinterface.github.io/shinydashboardPlus/reference/flipBox.html}
}
