% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_box.R
\name{flip_button_back}
\alias{flip_button_back}
\title{flip_button_back}
\usage{
flip_button_back(id, text)
}
\arguments{
\item{id}{the flipbox id}

\item{text}{the button text}
}
\description{
flip_button_back
}
