% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_box.R
\name{flip_button_front}
\alias{flip_button_front}
\title{flip_button_front}
\usage{
flip_button_front(id, text)
}
\arguments{
\item{id}{the flipbox id}

\item{text}{the button text}
}
\description{
flip_button_front
}
