% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipBox.R
\name{flipBoxDemo}
\alias{flipBoxDemo}
\title{flip_box_demo}
\usage{
flipBoxDemo()
}
\value{
shinyApp
}
\description{
runs an internal flip_box_demo application.
}
\examples{
\dontrun{
library(shinydashboardPlus)
flipBoxDemo()
}
}
