% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipBox.R
\name{flipButtonFront}
\alias{flipButtonFront}
\title{flipButtonFront}
\usage{
flipButtonFront(id, text)
}
\arguments{
\item{id}{the \code{flipbox} id}

\item{text}{the button text}
}
\description{
flipButtonFront
}
