% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced-boxes.R
\name{gradientBox}
\alias{gradientBox}
\title{AdminLTE2 gradient box}
\usage{
gradientBox(..., title = NULL, icon = NULL, gradientColor = NULL,
  boxToolSize = "sm", footer = NULL, width = 6, height = NULL,
  collapsible = TRUE, closable = FALSE)
}
\arguments{
\item{...}{body content.}

\item{title}{box title.}

\item{icon}{header icon, if any. Should be written like "fa fa-times".}

\item{gradientColor}{color of the box: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}

\item{boxToolSize}{size of the toolbox: choose among "xs", "sm", "md", "lg".}

\item{footer}{the box footer.}

\item{width}{box width (between 1 and 12).}

\item{height}{box height.}

\item{collapsible}{If TRUE, display a button in the upper right that allows the user to collapse the box.}

\item{closable}{If TRUE, display a button in the upper right that allows the user to close the box.}
}
\description{
Create an enhanced box
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     gradientBox(
      title = "My gradient Box",
      icon = "fa fa-th",
      gradientColor = "teal", 
      boxToolSize = "sm", 
      footer = fluidRow(
        column(
          width = 6,
          descriptionBlock(
            number = "17\%", 
            number_color = "green", 
            number_icon = "fa fa-caret-up",
            header = "$35,210.43", 
            text = "TOTAL REVENUE", 
            right_border = TRUE,
            margin_bottom = FALSE
          )
        ),
        column(
          width = 6,
          descriptionBlock(
            number = "18\%", 
            number_color = "red", 
            number_icon = "fa fa-caret-down",
            header = "1200", 
            text = "GOAL COMPLETION", 
            right_border = FALSE,
            margin_bottom = FALSE
          )
        )
      ),
      "This is a gradient box"
      )
    ),
    title = "Description Blocks"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
