% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-state.R
\name{loadingState}
\alias{loadingState}
\title{AdminLTE2 loading state element}
\usage{
loadingState()
}
\description{
When a section is still work in progress or a computation is running
}
\note{
Loading state can be programmatically used when a conputation is running for instance.
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "loading spinner",
      loadingState()
      )
    ),
    title = "Loading State"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
