% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced-boxes.R
\name{mailForm}
\alias{mailForm}
\title{AdminLTE2 mail form}
\usage{
mailForm(..., mailto = "#")
}
\arguments{
\item{...}{message text.}

\item{mailto}{person who should receive the mail.}
}
\description{
Create a mail form
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Mail box demo",
      mailForm(mailto = "dgranjon@ymail.com")
     )
    ),
    title = "mailForm"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
