% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{menuIcon}
\alias{menuIcon}
\title{AdminLTE2 menu icon item}
\usage{
menuIcon(name, color)
}
\arguments{
\item{name}{Icon name.}

\item{color}{background color: see here for a list of valid colors 
\url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create a nice menu icon to insert in a rightsidebarMenuItem.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
