% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{menuInfo}
\alias{menuInfo}
\title{AdminLTE2 menu info item}
\usage{
menuInfo(title = NULL, description = NULL)
}
\arguments{
\item{title}{Menu title.}

\item{description}{Extra informations.}
}
\description{
Create a nice menu info to insert in a rightsidebarMenuItem.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
