% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{messageItem}
\alias{messageItem}
\title{Custom messageItem}
\usage{
messageItem(
  inputId = NULL,
  from,
  message,
  icon = shiny::icon("user"),
  time = NULL,
  href = NULL
)
}
\arguments{
\item{inputId}{If not NULL, this item behaves like an action button.}

\item{from}{Who the message is from.}

\item{message}{Text of the message.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{time}{String representing the time the message was sent. Any string may
be used. For example, it could be a relative date/time like "5 minutes",
"today", or "12:30pm yesterday", or an absolute time, like "2014-12-01 13:45".
If NULL, no time will be displayed.}

\item{href}{An optional URL to link to.}
}
\description{
Custom messageItem
}
