% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preloader.R
\name{preloader}
\alias{preloader}
\title{AdminLTE2 preloader}
\usage{
preloader()
}
\description{
This creates a preloader.
}
\note{
This function is only for internal use of shinydashboardPlus.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
