% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{productListItem}
\alias{productListItem}
\title{AdminLTE2 product item}
\usage{
productListItem(..., src = NULL, title = NULL, subtitle = NULL, color = NULL)
}
\arguments{
\item{...}{product description.}

\item{src}{image url, if any.}

\item{title}{product name.}

\item{subtitle}{product price.}

\item{color}{price color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create a product item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
