% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeaderPlus.R
\name{renderUser}
\alias{renderUser}
\title{Create dynamic user output (server side)}
\usage{
renderUser(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link{HTML}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Create dynamic user output (server side)
}
\seealso{
\code{\link{userOutput}} for the corresponding client side function
  and examples.

Other user outputs: \code{\link{userOutput}}
}
\concept{user outputs}
