% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{rightSidebar}
\alias{rightSidebar}
\title{AdminLTE2 dashboard right sidebar}
\usage{
rightSidebar()
}
\description{
This creates a right sidebar.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
# Basic dashboard page template
library(shiny)
shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(
     enable_rightsidebar = TRUE,
     rightSidebarIcon = "gears"
    ),
    sidebar = dashboardSidebar(),
    body = dashboardBody(),
    rightsidebar = rightSidebar(),
    title = "Right Sidebar"
  ),
  server = function(input, output) { }
)
}
}
