% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{rightSidebarMenuItem}
\alias{rightSidebarMenuItem}
\title{AdminLTE2 right sidebar menu item}
\usage{
rightSidebarMenuItem(icon, info)
}
\arguments{
\item{icon}{Slot for menuIcon.}

\item{info}{Slot for menuInfo.}
}
\description{
Item to insert in a rightsidebarMenu.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
