% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{rightSidebarTabList}
\alias{rightSidebarTabList}
\title{AdminLTE2 right sidebar tab list}
\usage{
rightSidebarTabList(...)
}
\arguments{
\item{...}{slot that takes all rightSidebarTabContent as input to automatically
generate the same number of items in the tab menu with corresponding icons,
ids, ...}
}
\description{
This creates a right sidebar tab list.
}
