% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{rigthSidebarPanel}
\alias{rigthSidebarPanel}
\title{AdminLTE2 wrapper for tab content}
\usage{
rigthSidebarPanel(...)
}
\arguments{
\item{...}{slot for rightSidebarTabContent.}
}
\description{
This creates a wrapper that will contain rightSidebarTabContent.
}
