% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboardPlus-package.r
\docType{package}
\name{shinydashboardPlus}
\alias{shinydashboardPlus}
\alias{shinydashboardPlus-package}
\title{shinydashboardPlus}
\description{
Extend 'shinydashboard' with 'AdminLTE2' components. 
             'AdminLTE2' is a free 'Bootstrap 3' dashboard template available
             at <https://adminlte.io>. Customize boxes, add timelines and a lot more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RinteRface/shinydashboardPlus}
  \item \url{https://rinterface.com/shiny/shinydashboardPlus/}
  \item Report bugs at \url{https://github.com/RinteRface/shinydashboardPlus/issues}
}

}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com}

Other contributors:
\itemize{
  \item  RinteRface [copyright holder]
  \item  Almasaeed Studio (AdminLTE2 theme for Bootstrap 3) [contributor, copyright holder]
  \item Guang Yang (ygdashboard original template) [contributor, copyright holder]
  \item Winston Chang (Functions from shinydashboard) [contributor, copyright holder]
  \item Victor Perrier \email{victor.perrier@dreamrs.fr} (improved the shinydashboardPlusGallery) [contributor]
}

}
