% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboardPlusGallery.R
\name{shinydashboardPlusGallery}
\alias{shinydashboardPlusGallery}
\title{Launch the shinydashboardPlus Gallery}
\usage{
shinydashboardPlusGallery()
}
\description{
A gallery of all components available in shinydashboardPlus.
}
\examples{

if (interactive()) {

 shinydashboardPlusGallery()

}
}
