% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skinSelector.R
\name{skinSelector}
\alias{skinSelector}
\title{AdminLTE2 skin selector}
\usage{
skinSelector(...)
}
\arguments{
\item{...}{slot for skinSelectorItem.}
}
\description{
This creates a skin selector element.
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
   ui = dashboardPagePlus(
     header = dashboardHeaderPlus(
      enable_rightsidebar = TRUE,
      rightSidebarIcon = "gears"
     ),
     sidebar = dashboardSidebar(),
     body = dashboardBody(),
     rightsidebar = rightSidebar(
      background = "dark",
       rightSidebarTabList(
        rightSidebarTabItem(
         id = 1,
         icon = "desktop"
        ),
        rightSidebarTabItem(
         id = 2
        ),
        rightSidebarTabItem(
         id = 3,
         icon = "paint-brush"
        )
       ),
       rigthSidebarPanel(
        rightSidebarTabContent(
         id = 1,
         title = "Tab 1",
         skinSelector(
          skinSelectorItem(color = "red"),
          skinSelectorItem(color = "green"),
          skinSelectorItem(color = "yellow-light")
         )
        ),
        rightSidebarTabContent(
         id = 2,
         title = "Tab 2",
         textInput("caption", "Caption", "Data Summary")
        ),
        rightSidebarTabContent(
         id = 3,
         title = "Tab 3",
         numericInput("obs", "Observations:", 10, min = 1, max = 100)
        )
       )
     ),
     title = "Right Sidebar"
   ),
   server = function(input, output) { }
 )
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
