% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skinSelector.R
\name{skinSelectorItem}
\alias{skinSelectorItem}
\title{AdminLTE2 skin selector item}
\usage{
skinSelectorItem(color)
}
\arguments{
\item{color}{Skin color: "blue", "black", "purple", "red", "green", "yellow" as well as
"blue-light", "black-light", "purple-light", "red-light", "green-light" and  "yellow-light".}
}
\description{
This creates an item to insert in a skin selector object.
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
