% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{timelineEnd}
\alias{timelineEnd}
\title{AdminLTE2 timeline ending point}
\usage{
timelineEnd(icon = shiny::icon("hourglass-end"), color = NULL)
}
\arguments{
\item{icon}{item icon. Expect \code{\link[shiny]{icon}}.}

\item{color}{item color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create a timeline ending point
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
