% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timelineItemMedia}
\alias{timelineItemMedia}
\title{AdminLTE2 timeline media item}
\usage{
timelineItemMedia(src = NULL)
}
\arguments{
\item{src}{media url or path.}
}
\description{
Create a timeline media item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
