% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{timelineLabel}
\alias{timelineLabel}
\title{AdminLTE2 timeline label}
\usage{
timelineLabel(..., color = NULL)
}
\arguments{
\item{...}{any element.}

\item{color}{label color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create a timeline label
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
