% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{todoListItem}
\alias{todoListItem}
\title{AdminLTE2 todo list item}
\usage{
todoListItem(..., checked = FALSE, label = NULL)
}
\arguments{
\item{...}{any element such as labels, ...}

\item{checked}{Whether the list item is checked or not.}

\item{label}{item label.}
}
\description{
Create a todo list item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
