% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{updateBoxPlus}
\alias{updateBoxPlus}
\title{Collapse a \link{boxPlus} tag.}
\usage{
updateBoxPlus(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Box to toggle.}

\item{session}{Shiny session object.}
}
\description{
Collapse a \link{boxPlus} tag.
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 ui <- dashboardPagePlus(
   dashboardHeaderPlus(),
   dashboardSidebar(),
   dashboardBody(
     tags$style("body { background-color: ghostwhite};"),
     
     br(),
     boxPlus(
       title = textOutput("box_state"),
       "Box body",
       inputId = "mybox",
       collapsible = TRUE,
       plotOutput("plot")
     ),
     actionButton("toggle_box", "Toggle Box", class = "bg-success")
   )
 )
 
 server <- function(input, output, session) {
   output$plot <- renderPlot({
     req(!input$mybox$collapsed)
     plot(rnorm(200))
   })
   
   output$box_state <- renderText({
     state <- if (input$mybox$collapsed) "collapsed" else "uncollapsed"
     paste("My box is", state)
   })
   
   observeEvent(input$toggle_box, {
     updateBoxPlus("mybox")
   })
   
 }
 
 shinyApp(ui, server)
}
}
