% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{updateBoxPlusSidebar}
\alias{updateBoxPlusSidebar}
\title{Programmatically toggle a \link{boxPlusSidebar}}
\usage{
updateBoxPlusSidebar(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Sidebar id.}

\item{session}{Shiny session object.}
}
\description{
Programmatically toggle a \link{boxPlusSidebar}
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPagePlus(
    header = dashboardHeader(),
    body = dashboardBody(
      boxPlus(
        title = "Update box sidebar", 
        closable = TRUE, 
        width = 12,
        height = "500px",
        solidHeader = FALSE, 
        collapsible = TRUE,
        actionButton("update", "Toggle card sidebar"),
        sidebar = boxPlusSidebar(
          inputId = "mycardsidebar",
          p("Sidebar Content")
        )
      )
    ),
    sidebar = dashboardSidebar()
  ),
  server = function(input, output, session) {
    observe(print(input$mycardsidebar))
    
    observeEvent(input$update, {
      updateBoxPlusSidebar(inputId = "mycardsidebar")
    })
    
  }
 )
}
}
