% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{updateBoxSidebar}
\alias{updateBoxSidebar}
\title{Programmatically toggle a \link{boxSidebar}}
\usage{
updateBoxSidebar(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Sidebar id.}

\item{session}{Shiny session object.}
}
\description{
Programmatically toggle a \link{boxSidebar}
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    header = dashboardHeader(),
    body = dashboardBody(
      box(
        title = "Update box sidebar", 
        closable = TRUE, 
        width = 12,
        height = "500px",
        solidHeader = FALSE, 
        collapsible = TRUE,
        actionButton("update", "Toggle card sidebar"),
        sidebar = boxSidebar(
          id = "mycardsidebar",
          p("Sidebar Content")
        )
      )
    ),
    sidebar = dashboardSidebar()
  ),
  server = function(input, output, session) {
    observe(print(input$mycardsidebar))
    
    observeEvent(input$update, {
      updateBoxSidebar("mycardsidebar")
    })
    
  }
 )
}
}
