% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{updateNavPills}
\alias{updateNavPills}
\title{Update navPills on the client}
\usage{
updateNavPills(id, selected, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{\link{navPills} unique id to target.}

\item{selected}{Index of the \link{navPillsItem} to select. Index is seen from the R side.}

\item{session}{Shiny session object.}
}
\description{
Update navPills on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     radioButtons("controller", "Controller", choices = c(1, 2, 3)),
     br(),
     box(
      title = "Nav Pills",
      status = "info",
      "Box Body",
      footer = navPills(
        inputId = "pills",
        navPillsItem(
          left = "Item 1", 
          color = "green",
          right = 10
        ),
        navPillsItem(
          left = "Item 2", 
          color = "red",
          icon = icon("angle-down"), 
          right = "10\%"
        ),
        navPillsItem(
          left = "Item 3", 
          color = "blue",
          icon = icon("angle-up"), 
          right = "30\%"
        )
      )
     )
    ),
    title = "Nav Pills"
  ),
  server = function(input, output, session) {
   observeEvent(input$controller, {
    updateNavPills(id = "pills", selected = input$controller)
   })
   observeEvent(input$pills, {
    showNotification(sprintf("You selected pill N° \%s", input$pills), type = "message")
   })
  }
 )
}
}
