% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{userBox}
\alias{userBox}
\title{AdminLTE2 user box}
\usage{
userBox(
  ...,
  title,
  subtitle = NULL,
  footer = NULL,
  color = NULL,
  width = 6,
  height = NULL,
  collapsible = TRUE,
  collapsed = FALSE,
  closable = FALSE,
  type = c(1, 2),
  image = NULL,
  backgroundImage = NULL,
  boxToolSize = "sm",
  footerPadding = TRUE
)
}
\arguments{
\item{...}{body content.}

\item{title}{box title.}

\item{subtitle}{box subtitle.}

\item{footer}{box footer.}

\item{color}{background color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.
See below:
\itemize{
 \item \code{light_blue (primary status)}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#3c8dbc")}.
 \item \code{red (danger status)}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#dd4b39")}.
 \item \code{green (success status)}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#00a65a")}.
 \item \code{aqua (info status)}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#00c0ef")}.
 \item \code{yellow (warning status)}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#f39c12")}.
 \item \code{blue}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#0073b7")}.
 \item \code{navy}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#001F3F")}.
 \item \code{teal}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#39CCCC")}.
 \item \code{olive}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#3D9970")}.
 \item \code{lime}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#01FF70")}.
 \item \code{orange}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#FF851B")}.
 \item \code{fuchsia}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#F012BE")}.
 \item \code{purple}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#605ca8")}.
 \item \code{maroon}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#D81B60")}.
 \item \code{black}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#111")}.
 \item \code{gray_lte}: \Sexpr[results=rd, stage=install]{shinydashboardPlus:::rd_color_tag("#d2d6de")}.
}}

\item{width}{box width (between 1 and 12).}

\item{height}{box height.}

\item{collapsible}{If TRUE, display a button in the upper right that allows the user to collapse the box.}

\item{collapsed}{If TRUE, start collapsed. This must be used with \code{collapsible=TRUE}.}

\item{closable}{If TRUE, display a button in the upper right that allows the user to close the box.}

\item{type}{User box type. Either 1 or 2. 1 corresponds to a centered user image,
while 2 is a left aligned user image.}

\item{image}{header image, if any (this is different of the background image).}

\item{backgroundImage}{image url, if any. Background needs to be TRUE.}

\item{boxToolSize}{size of the toolbox: choose among "xs", "sm", "md", "lg".}

\item{footerPadding}{TRUE by default: whether the footer has margin or not.}
}
\description{
Create user box
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     userBox(
      title = "Nadia Carmichael",
      subtitle = "lead Developer",
      type = 2,
      image = "https://adminlte.io/themes/AdminLTE/dist/img/user7-128x128.jpg",
      color = "yellow",
      "Some text here!",
      footer = "The footer here!"
     ),
     userBox(
      title = "Alexander Pierce",
      subtitle = "Founder & CEO",
      type = 1,
      image = "https://adminlte.io/themes/AdminLTE/dist/img/user1-128x128.jpg",
      color = "aqua",
      closable = TRUE,
      "Some text here!",
      footer = "The footer here!"
     ),
     userBox(
      title = "Elizabeth Pierce",
      subtitle = "Web Designer",
      image = "https://adminlte.io/themes/AdminLTE/dist/img/user3-128x128.jpg",
      backgroundImage = "https://cdn.statically.io/img/wallpaperaccess.com/full/1119564.jpg",
      closable = TRUE,
      "Some text here!",
      footer = "The footer here!"
     )
    ),
    title = "UserBox"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
