% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userListItem}
\alias{userListItem}
\title{AdminLTE2 user list item}
\usage{
userListItem(image = NULL, title = NULL, subtitle = NULL)
}
\arguments{
\item{image}{image url or path.}

\item{title}{Item title.}

\item{subtitle}{Item subitle.}
}
\description{
Create a user list item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
