% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeaderPlus.R
\name{userOutput}
\alias{userOutput}
\title{Create a dynamic user output (client side)}
\usage{
userOutput(id, tag = shiny::tags$li)
}
\arguments{
\item{id}{Output variable name.}

\item{tag}{A tag function, like \code{tags$li} or \code{tags$ul}.}
}
\description{
This can be used as a placeholder for dynamically-generated \code{\link{dashboardUser}}.
}
\seealso{
\code{\link{renderUser}} for the corresponding server side function
  and examples.

Other user outputs: 
\code{\link{renderUser}()}
}
\concept{user outputs}
