% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-post.R
\name{userPostMedia}
\alias{userPostMedia}
\title{AdminLTE2 user post media}
\usage{
userPostMedia(src = NULL)
}
\arguments{
\item{src}{image path or url ...}
}
\description{
Create a user post media (image)
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
