% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-post.R
\name{userPostToolItem}
\alias{userPostToolItem}
\title{AdminLTE2 user post tool item}
\usage{
userPostToolItem(..., side = "left")
}
\arguments{
\item{...}{tool content such as label, button, ...}

\item{side}{tool item orientation: "left" of "right", "left" by default.}
}
\description{
Create a user post tool item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
