% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-post.R
\name{userPostToolItemList}
\alias{userPostToolItemList}
\title{AdminLTE2 user post tool item container}
\usage{
userPostToolItemList(...)
}
\arguments{
\item{...}{slot for userPostToolItem.}
}
\description{
Create a user post tool item container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
