% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{validStatusesPlus}
\alias{validStatusesPlus}
\title{Valid statuses extra}
\description{
These status strings correspond to colors as defined in adminLTE extra colors.
Although the colors can vary depending on the particular CSS selector, they
generally appear as follows:
}
\details{
\itemize{
  \item \code{primary} Blue (sometimes dark blue)
  \item \code{success} Green
  \item \code{info} Blue
  \item \code{warning} Orange
  \item \code{danger} Red
  \item \code{navy} Dark Grey/Blue
  \item \code{teal} Blue/Green
  \item \code{orange} Orange
  \item \code{purple} Purple
  \item \code{maroon} Pink
  \item \code{black} Black
}
}
\keyword{internal}
