% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{accordionItem}
\alias{accordionItem}
\title{AdminLTE2 accordion item}
\usage{
accordionItem(..., id, title = NULL, color = NULL, collapsed = TRUE)
}
\arguments{
\item{...}{text to write in the item.}

\item{id}{unique item id.}

\item{title}{item title.}

\item{color}{item color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}

\item{collapsed}{Whether to expand or collapse the item. TRUE by default. Set it to FALSE if you want to expand it.}
}
\description{
Create an accordion item to put inside an accordion container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
