% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockQuote.R
\name{blockQuote}
\alias{blockQuote}
\title{AdminLTE2 block quote}
\usage{
blockQuote(..., side = "left")
}
\arguments{
\item{...}{any element.}

\item{side}{blockauote orientation. "left" by default, can be set to "right".}
}
\description{
If you want to quote text
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "BlockQuote example",
      blockQuote("I quote some text here!")
     )
    ),
    title = "blockQuote"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
