% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{dashboardBadge}
\alias{dashboardBadge}
\title{AdminLTE2 badge}
\usage{
dashboardBadge(..., color)
}
\arguments{
\item{...}{Any html text element.}

\item{color}{label color.}
}
\description{
Create a badge.
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     dashboardBadge("Badge 1", color = "blue"),
     actionButton(
      inputId = "badge", 
      label = "Hello", 
      icon = NULL, 
      width = NULL, 
      dashboardBadge(1, color = "orange")
     )
    )
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
