% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{dashboardControlbar}
\alias{dashboardControlbar}
\title{AdminLTE2 dashboard right sidebar}
\usage{
dashboardControlbar(
  ...,
  inputId = NULL,
  skin = "dark",
  disable = FALSE,
  collapsed = TRUE,
  overlay = TRUE,
  width = 230,
  .items = NULL
)
}
\arguments{
\item{...}{slot for rightSidebarTabContent. Not compatible with .items.}

\item{inputId}{To acces the current state of the controlbar. Open is TRUE, closed
is FALSE. NULL by default.}

\item{skin}{background color: "dark" or "light".}

\item{disable}{If \code{TRUE}, the sidebar will be disabled.}

\item{collapsed}{Whether the control bar on the right side is collapsed or not at start. TRUE by default.}

\item{overlay}{Whether the sidebar covers the content when expanded. Default to TRUE.}

\item{width}{Sidebar width in pixels. Numeric value expected. 230 by default.}

\item{.items}{Pass element here if you do not want to embed them in panels. Not compatible with ...}
}
\description{
This creates a right sidebar.
}
\note{
Until a maximum of 5 rightSidebarTabContent inside! AdminLTE 2 does not
support more panels.
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
   ui = dashboardPagePlus(
     header = dashboardHeaderPlus(
      enable_rightsidebar = TRUE,
      rightSidebarIcon = "gears"
     ),
     sidebar = dashboardSidebar(),
     body = dashboardBody(),
     controlbar = dashboardControlbar(
      skin = "dark",
       rightSidebarTabContent(
        id = 1,
        icon = "desktop",
        title = "Tab 1",
        active = TRUE,
        sliderInput(
         "obs", 
         "Number of observations:",
          min = 0, max = 1000, value = 500
         )
       ),
       rightSidebarTabContent(
        id = 2,
        title = "Tab 2",
        textInput("caption", "Caption", "Data Summary")
       ),
       rightSidebarTabContent(
        id = 3,
        title = "Tab 3",
        icon = "paint-brush",
        numericInput("obs", "Observations:", 10, min = 1, max = 100)
       )
     ),
     title = "Right Sidebar"
   ),
   server = function(input, output) { }
 )
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
