% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{dashboardControlbar}
\alias{dashboardControlbar}
\title{AdminLTE2 dashboard right sidebar}
\usage{
dashboardControlbar(
  ...,
  inputId = NULL,
  skin = "dark",
  disable = FALSE,
  collapsed = TRUE,
  overlay = TRUE,
  width = 230,
  .items = NULL
)
}
\arguments{
\item{...}{slot for rightSidebarTabContent. Not compatible with .items.}

\item{inputId}{To acces the current state of the controlbar. Open is TRUE, closed
is FALSE. NULL by default.}

\item{skin}{background color: "dark" or "light".}

\item{disable}{If \code{TRUE}, the sidebar will be disabled.}

\item{collapsed}{Whether the control bar on the right side is collapsed or not at start. TRUE by default.}

\item{overlay}{Whether the sidebar covers the content when expanded. Default to TRUE.}

\item{width}{Sidebar width in pixels. Numeric value expected. 230 by default.}

\item{.items}{Pass element here if you do not want to embed them in panels. Not compatible with ...}
}
\description{
This creates a right sidebar.
}
\note{
Until a maximum of 5 rightSidebarTabContent inside! AdminLTE 2 does not
support more panels.
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 shinyApp(
   ui = dashboardPage(
     header = dashboardHeader(),
     sidebar = dashboardSidebar(),
     body = dashboardBody(),
     controlbar = dashboardControlbar(
      skin = "dark",
      controlbarMenu(
       id = "menu",
       controlbarItem(
        "Tab 1",
        "Welcome to tab 1"
       ),
       controlbarItem(
        "Tab 2",
        "Welcome to tab 2"
       )
      )
     ),
     title = "Right Sidebar"
   ),
   server = function(input, output) { }
 )
}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
