% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{dashboardFooter}
\alias{dashboardFooter}
\title{Dashboard Footer}
\usage{
dashboardFooter(left_text = NULL, right_text = NULL)
}
\arguments{
\item{left_text}{Left text.}

\item{right_text}{Right text.}
}
\description{
This creates a dashboard footer
}
\examples{
if (interactive()) {
library(shiny)
library(shinydashboard)
library(shinydashboardPlus)

shinyApp(
  ui = dashboardPagePlus(
    header = dashboardHeaderPlus(
     enable_rightsidebar = TRUE,
     rightSidebarIcon = "gears"
    ),
    sidebar = dashboardSidebar(),
    body = dashboardBody(),
    rightsidebar = rightSidebar(),
    footer = dashboardFooter(
     left_text = "By Divad Nojnarg",
     right_text = "Zurich, 2019"
    ),
    title = "DashboardPage"
  ),
  server = function(input, output) { }
)
}
}
