% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardLabel.R
\name{dashboardLabel}
\alias{dashboardLabel}
\title{AdminLTE2 label}
\usage{
dashboardLabel(..., status = "primary")
}
\arguments{
\item{...}{any text.}

\item{status}{label status: "danger", "success", "info", "primary", "warning".}
}
\description{
Create a label
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     dashboardLabel("Label 1", status = "info")
    )
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
