% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{dashboardSidebar}
\alias{dashboardSidebar}
\title{Create a dashboard sidebar.}
\usage{
dashboardSidebar(
  ...,
  inputId = NULL,
  disable = FALSE,
  width = NULL,
  collapsed = FALSE,
  minified = FALSE
)
}
\arguments{
\item{...}{Items to put in the sidebar.}

\item{inputId}{Sidebar input id. Returns the current state of the sidebar.}

\item{disable}{If \code{TRUE}, the sidebar will be disabled.}

\item{width}{The width of the sidebar. This must either be a number which
specifies the width in pixels, or a string that specifies the width in CSS
units.}

\item{collapsed}{If \code{TRUE}, the sidebar will be collapsed on app startup.}

\item{minified}{Whether to slightly close the sidebar but still show item icons. Default
to FALSE.}
}
\description{
A dashboard sidebar typically contains a \code{\link{sidebarMenu}}, although
it may also contain a \code{\link{sidebarSearchForm}}, or other Shiny inputs.
}
