% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{dropdownBlock}
\alias{dropdownBlock}
\title{Create a dropdown block to place in a dashboard header}
\usage{
dropdownBlock(..., id, icon = NULL, title = NULL, badgeStatus = "danger")
}
\arguments{
\item{...}{Items to put in the menu.}

\item{id}{Dropdown block id.}

\item{icon}{An icon to display in the header. Expect \code{\link[shiny]{icon}}.}

\item{title}{Dropdown block title.}

\item{badgeStatus}{Dropdown badge status.}
}
\description{
Create a dropdown block to place in a dashboard header
}
\seealso{
\code{\link{dashboardHeader}} for example usage.
}
