% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipBox.R
\name{flip_button_front}
\alias{flip_button_front}
\alias{flip_button_back}
\alias{flip_box}
\title{flip_box}
\usage{
flip_button_front(id, text)

flip_button_back(id, text)

flip_box(
  id,
  front_content = htmltools::h1("Hi From the Front"),
  back_content = tags$div(style = "background-color: #FFF; height: 380px;",
    back_content)
)
}
\arguments{
\item{id}{the flipbox id}

\item{text}{the button text}

\item{front_content}{ui for the front of the flip box}

\item{back_content}{ui for the back of the flip box}
}
\value{
HTML tagList
}
\description{
UI for the flip_box - based off of \href{https://rinterface.github.io/shinydashboardPlus/}{shinydashboardplus::flipBox}
}
\section{Functions}{
\itemize{
\item \code{flip_button_front}: flip_button_front button to flip to the front of the \code{flip_box}.

\item \code{flip_button_back}: flip_button_back button to flip to the back of the \code{flip_box}.
}}

\section{Warning}{
 this function is experimental.
}

\examples{
library(shiny)
library(shinydashboard)
library(shinydashboardPlus)
shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
      box(
        width = 12,
        style = paste0("background-color: #000; color: #FFF;"),
        fluidRow(
          column(
            width = 12,
            flip_box(
              id = "id",
              front_content = img(
                src = "https://image.flaticon.com/icons/svg/149/149076.svg",
                width = "100\%",
                height = "300px",
                style = "border: 1px solid black"
              ),
              back_content = div(
                style = "border: 1px solid black; height: 300px; background-color: #FFFFFF; color: #000000;",
                h1(
                  class = "text-center",
                  "Details...."
                ),
                p("More information....")
              )
            )
          )
        ),
        
        fluidRow(
          column(
            width = 12,
            br(),
            div(
              id = "go_to_back",
              class = "pull_right",
              style = "display: inline-block; margin-right: 5px;",
              flip_button_front("id", "More Info")
            ),
            div(
              id = "go_to_front",
              class = "pull-right",
              style = "display: inline-block; margin-right: 5px;",
              flip_button_back("id", "Back to Front")
            )
          )
        )
      )
    )
  ),
  
  server = function(input, output) {
    
  }
)
   

}
\references{
\url{https://rinterface.github.io/shinydashboardPlus/reference/flipBox.html}
}
