% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R, R/flipBox.R
\name{flipBox}
\alias{flipBox}
\title{AdminLTE2 flipping box}
\usage{
flipBox(
  id,
  front_content,
  back_content,
  front_text = "More Info",
  back_text = "Back to Front"
)

flipBox(
  id,
  front_content,
  back_content,
  front_text = "More Info",
  back_text = "Back to Front"
)
}
\arguments{
\item{id}{the \code{flipBox} id}

\item{front_content}{ui for the front of the flip box}

\item{back_content}{ui for the back of the flip box}

\item{front_text}{text for front button}

\item{back_text}{text for back button}

\item{...}{front body content.}

\item{front_title}{Box front title.}

\item{back_title}{Box back title.}

\item{front_btn_text}{Front button text.}

\item{back_btn_text}{Back button text.}

\item{header_img}{Header background image url or path.}

\item{main_img}{Main image url or path (for instance profile image).}

\item{width}{box width (between 1 and 12). 6 by default.}
}
\value{
HTML tagList
}
\description{
Create a flipping box

Create a flipping box based off of the AdminLTE flipBox.
}
\details{
To access the state of the flipbox use the input alias \code{input$<id>_value}.
  For example, if your flipBox's id is myawesomeflipbox, you can access its state via
  \code{input$myawesomeflipbox_value} where a 0 corresponds to the front, 1 the back, and NULL 
  indicates the input has not been set yet (i.e. the card has not flipped yet).
}
\section{Warning}{
 this function is experimental.
}

\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
   ui = dashboardPage(
     dashboardHeader(),
     dashboardSidebar(),
     dashboardBody(
       setShadow("card"),
       fluidRow(
         column(
           width = 6,
           align = "center",
           flipBox(
             id = 1,
             main_img = "https://image.flaticon.com/icons/svg/149/149076.svg",
             header_img = "https://image.flaticon.com/icons/svg/119/119595.svg",
             front_title = "John Doe",
             back_title = "About John",
             "Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
             sed do eiusmod tempor incididunt ut labore et dolore magna 
             aliqua. Ut enim ad minim veniam, quis nostrud exercitation 
             ullamco laboris nisi ut aliquip ex ea commodo consequat. 
             Duis aute irure dolor in reprehenderit in voluptate velit 
             esse cillum dolore eu fugiat nulla pariatur. Excepteur sint 
             occaecat cupidatat non proident, sunt in culpa qui officia 
             deserunt mollit anim id est laborum",
             fluidRow(
               dashboardLabel("Label 1", status = "info"),
               dashboardLabel("Label 2", status = "success"),
               dashboardLabel("Label 3", status = "warning"),
               dashboardLabel("Label 4", status = "primary"),
               dashboardLabel("Label 5", status = "danger")
             ),
             hr(),
             fluidRow(
               column(
                 width = 6,
                 align = "center",
                 starBlock(grade = 5),
                 starBlock(grade = 5, color = "olive"),
                 starBlock(grade = 1, color = "maroon"),
                 starBlock(grade = 3, color = "teal")
               ),
               column(
                 width = 6,
                 align = "center",
                 appButton(
                   url = "https://google.com",
                   label = "Users",
                   icon = "fa fa-users",
                   enable_badge = TRUE,
                   badgeColor = "purple",
                   badgeLabel = 891
                 ),
                 appButton(
                   label = "Edit",
                   icon = "fa fa-edit",
                   enable_badge = FALSE,
                   badgeColor = NULL,
                   badgeLabel = NULL
                 )
               )
             ),
             back_content = tagList(
               column(
                 width = 12,
                 align = "center",
                 sliderInput(
                   "obs", 
                   "Number of observations:",
                   min = 0, 
                   max = 1000, 
                   value = 500
                 )
               ),
               plotOutput("distPlot")
             )
           )
       ),
       column(
         width = 6,
         align = "center",
         flipBox(
           id = 2,
           main_img = "https://image.flaticon.com/icons/svg/149/149073.svg",
           header_img = "https://image.flaticon.com/icons/svg/119/119598.svg",
           front_title = "Johanna Doe",
           back_title = "About Johanna",
           fluidRow(
             column(
               width = 6,
               align = "center",
               boxPad(
                 color = "green",
                 descriptionBlock(
                   header = "8390",
                   text = "VISITS",
                   right_border = FALSE,
                   margin_bottom = TRUE
                 ),
                 descriptionBlock(
                   header = "30\%",
                   text = "REFERRALS",
                   right_border = FALSE,
                   margin_bottom = TRUE
                 ),
                 descriptionBlock(
                   header = "70\%",
                   text = "ORGANIC",
                   right_border = FALSE,
                   margin_bottom = FALSE
                 )
               )
             ),
             column(
               width = 6,
               align = "center",
               "Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
               sed do eiusmod tempor.",
               br(),
               verticalProgress(
                 value = 10,
                 striped = TRUE,
                 active = TRUE
               ),
               verticalProgress(
                 value = 50,
                 active = TRUE,
                 status = "warning",
                 size = "xs"
               ),
               verticalProgress(
                 value = 20,
                 status = "danger",
                 size = "sm",
                 height = "60\%"
               )
             )
           ),
           back_content = tagList(
             column(
               width = 12,
               align = "center",
               radioButtons(
                 "dist", 
                 "Distribution type:",
                 c("Normal" = "norm",
                   "Uniform" = "unif",
                   "Log-normal" = "lnorm",
                   "Exponential" = "exp"
                 )
               )
             ),
             plotOutput("plot")
           )
       )
       )
     )
   ),
   title = "flipBox"
   ),
   server = function(input, output) {
     output$distPlot <- renderPlot({
       hist(rnorm(input$obs))
     })
     output$plot <- renderPlot({
       dist <- switch(input$dist,
                      norm = rnorm,
                      unif = runif,
                      lnorm = rlnorm,
                      exp = rexp,
                      rnorm)
       
       hist(dist(500))
     })
   }
  )
}

library(shiny)
library(shinydashboard)
library(shinydashboardPlus)
shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
      flip_box(
        id = "id",
        front_content = div(
          class = "text-center",
          img(
            src = "https://image.flaticon.com/icons/svg/149/149076.svg",
            height = "300px",
            width = "100\%"
          )
        ),
        back_content = div(
          class = "text-center",
          height = "300px",
          width = "100\%",
          h1("Details...."),
          p("More information....")
        )
      )
    )
  ),
  
  server = function(input, output, session) {
  }
)
   
}
\references{
\url{https://rinterface.github.io/shinydashboardPlus/reference/flipBox.html}
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
