% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipBox.R
\name{flipBoxDemo}
\alias{flipBoxDemo}
\title{flipBoxDemo}
\usage{
flipBoxDemo()
}
\value{
shinyApp
}
\description{
runs an internal flipBoxDemo application.
}
\examples{
\dontrun{
library(shinydashboardPlus)
flipBoxDemo()
}
}
