% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{navPills}
\alias{navPills}
\title{AdminLTE2 nav pill container}
\usage{
navPills(..., id = NULL)
}
\arguments{
\item{...}{slot for navPillsItem.}

\item{id}{Item unique id. Returns the R index of the currently selected item. 
This is different from the JavaScript index.}
}
\description{
Create a container for nav elements
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Nav Pills",
      status = "info",
      "Box Body",
      footer = navPills(
        id = "pillItem",
        navPillsItem(
          left = "Item 1", 
          color = "green",
          right = 10
        ),
        navPillsItem(
          left = "Item 2", 
          color = "red",
          icon = icon("angle-down"), 
          right = "10\%",
          active = TRUE
        )
      )
     )
    ),
    title = "Nav Pills"
  ),
  server = function(input, output) {
   observeEvent(input$pillItem, {
    showNotification(sprintf("You clicked on pill N° \%s", input$pillItem), type = "message")
   })
  }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
