% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{navPills}
\alias{navPills}
\title{AdminLTE2 nav pill container}
\usage{
navPills(...)
}
\arguments{
\item{...}{slot for navPillsItem.}
}
\description{
Create a container for nav elements
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Nav Pills",
      status = "info",
      "Box Body",
      footer = navPills(
        navPillsItem(
          pillName = "Item 1", 
          pillColor = "green",
          pillIcon = NULL, 
          pillText = 10
        ),
        navPillsItem(
          pillName = "Item 2", 
          pillColor = "red",
          pillIcon = "fa fa-angle-down", 
          pillText = "10\%",
          active = TRUE
        )
      )
     )
    ),
    title = "Nav Pills"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
