% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{navPillsItem}
\alias{navPillsItem}
\title{AdminLTE2 nav pill item}
\usage{
navPillsItem(
  left = NULL,
  right = NULL,
  color = NULL,
  icon = NULL,
  active = FALSE
)
}
\arguments{
\item{left}{pill left text.}

\item{right}{pill right text.}

\item{color}{pill color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}

\item{icon}{pill icon, if any.}

\item{active}{Whether the item is active or not. FALSE by default.}
}
\description{
Create a nav pill item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
