% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{navPillsItem}
\alias{navPillsItem}
\title{AdminLTE2 nav pill item}
\usage{
navPillsItem(pillName = NULL, pillColor = NULL, pillIcon = NULL,
  pillText = NULL, active = FALSE)
}
\arguments{
\item{pillName}{pill name.}

\item{pillColor}{pill color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}

\item{pillIcon}{pill icon, if any. Should be written like "fa fa-times".}

\item{pillText}{pill text. Can also be a number.}

\item{active}{Whether the item is active or not. FALSE by default.}
}
\description{
Create a nav pill item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
