% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product-list.R
\name{productList}
\alias{productList}
\title{AdminLTE2 product list container}
\usage{
productList(...)
}
\arguments{
\item{...}{slot for productListItem.}
}
\description{
Create a container product list elements
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Product List",
      status = "primary",
      productList(
        productListItem(
          src = "https://image.darty.com/hifi_video/televiseurs-led/grand_ecran_led/samsung_ue55mu9005_4k_uhd_s1703224306074A_140323143.jpg", 
          productTitle = "Samsung TV", 
          productPrice = "$1800", 
          priceColor = "warning",
          "This is an amazing TV, but I don't like TV!"
        ),
        productListItem(
          src = "http://icon-park.com/imagefiles/imac.png", 
          productTitle = "Imac 27", 
          productPrice = "$2400", 
          priceColor = "danger",
          "This is were I spend most of my time!"
        )
      )
     )
    ),
    title = "Description Blocks"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
