% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product-list.R
\name{productListItem}
\alias{productListItem}
\title{AdminLTE2 product item}
\usage{
productListItem(..., src = NULL, productTitle = NULL, productPrice = NULL,
  priceColor = NULL)
}
\arguments{
\item{...}{product description.}

\item{src}{image url, if any.}

\item{productTitle}{product name.}

\item{productPrice}{product price.}

\item{priceColor}{price color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create a product item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
