% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{rightSidebarTabContent}
\alias{rightSidebarTabContent}
\title{AdminLTE2 tab content}
\usage{
rightSidebarTabContent(..., id, title = NULL)
}
\arguments{
\item{...}{any element such as sliderInput, ...}

\item{id}{should correspond to the id given in rightSidebarTabItem.}

\item{title}{content title.}
}
\description{
This creates a wrapper that will contain rightSidebarTabContent.
}
