% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{rightSidebarTabItem}
\alias{rightSidebarTabItem}
\title{AdminLTE2 right sidebar tab item}
\usage{
rightSidebarTabItem(id, icon, active)
}
\arguments{
\item{id}{unique item id.}

\item{icon}{tab icon.}

\item{active}{Whether the tab item is active or not.}
}
\description{
This creates a right sidebar tab item to be inserted in a rightSidebarTabList.
}
