% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightSidebar.R
\name{rightSidebarTabList}
\alias{rightSidebarTabList}
\title{AdminLTE2 right sidebar tab list}
\usage{
rightSidebarTabList(...)
}
\arguments{
\item{...}{slot for rightSidebarTabItem.}
}
\description{
This creates a right sidebar tab list.
}
\note{
Until a maximum of 5 rightSidebarTabItem!
}
