% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{socialBox}
\alias{socialBox}
\title{AdminLTE2 social box}
\usage{
socialBox(
  ...,
  image,
  title,
  subtitle = NULL,
  comments = NULL,
  footer = NULL,
  width = 6,
  height = NULL,
  collapsible = TRUE,
  closable = FALSE,
  footerPadding = TRUE
)
}
\arguments{
\item{...}{body content. May include \link{attachmentBlock} for instance.}

\item{image}{header image, if any.}

\item{title}{box title.}

\item{subtitle}{box subtitle.}

\item{comments}{slot for \link{boxComment}.}

\item{footer}{box footer, if any.}

\item{width}{box width (between 1 and 12).}

\item{height}{box height.}

\item{collapsible}{If TRUE, display a button in the upper right that allows the user to collapse the box.}

\item{closable}{If TRUE, display a button in the upper right that allows the user to close the box.}

\item{footerPadding}{TRUE by default: whether the footer has margin or not.}
}
\description{
Create social box
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 library(shinydashboardPlus)
 
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     socialBox(
      title = "Social Box",
      subtitle = "example-01.05.2018",
      image = "https://adminlte.io/themes/AdminLTE/dist/img/user4-128x128.jpg",
      "Some text here!",
      attachmentBlock(
       image = "https://adminlte.io/themes/AdminLTE/dist/img/photo1.png",
       title = "Test",
       title_url = "https://google.com",
       "This is the content"
      ),
      comments = tagList(
       lapply(X = 1:10, FUN = function(i) {
        boxComment(
          image = "https://adminlte.io/themes/AdminLTE/dist/img/user3-128x128.jpg",
          title = paste("Comment", i),
          date = "01.05.2018",
          paste0("The ", i, "-th comment")
        )
       })
      ),
      footer = "The footer here!"
     )
    ),
    title = "socialBox"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
