% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{socialButton}
\alias{socialButton}
\title{AdminLTE2 social button}
\usage{
socialButton(url, type = NULL)
}
\arguments{
\item{url}{if the button should redirect somewhere.}

\item{type}{social network name: see here for valid names \url{https://adminlte.io/themes/AdminLTE/pages/UI/buttons.html}.}
}
\description{
Create a social button
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboard)
 shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody(
     box(
      title = "Social Buttons",
      status = NULL,
      socialButton(
        url = "http://dropbox.com",
        type = "dropbox"
      ),
      socialButton(
        url = "http://github.com",
        type = "github"
      )
     )
    ),
    title = "Description Blocks"
  ),
  server = function(input, output) { }
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
