% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{timelineItemMedia}
\alias{timelineItemMedia}
\title{AdminLTE2 timeline media item}
\usage{
timelineItemMedia(src = NULL, height = NULL, width = NULL)
}
\arguments{
\item{src}{media url or path.}

\item{height}{media height in pixels.}

\item{width}{media width in pixels.}
}
\description{
Create a timeline media item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
