% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timelineStart}
\alias{timelineStart}
\title{AdminLTE2 timeline starting point}
\usage{
timelineStart(icon = "clock-o", color = NULL)
}
\arguments{
\item{icon}{item icon such as "clock-o", "times", ...}

\item{color}{item color: see here for a list of valid colors \url{https://adminlte.io/themes/AdminLTE/pages/UI/general.html}.}
}
\description{
Create a timeline starting point
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
