% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{updateControlbar}
\alias{updateControlbar}
\title{Function to programmatically toggle the state of the controlbar}
\usage{
updateControlbar(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Controlbar id.}

\item{session}{Shiny session object.}
}
\description{
Function to programmatically toggle the state of the controlbar
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinydashboardPlus)
 
 shinyApp(
   ui = dashboardPagePlus(
     header = dashboardHeaderPlus(),
     sidebar = dashboardSidebar(),
     body = dashboardBody(
       actionButton(inputId = "controlbarToggle", label = "Toggle Controlbar")
     ),
     controlbar = dashboardControlbar(inputId = "controlbar")
   ),
   server = function(input, output, session) {
     
     observeEvent(input$controlbar, {
       if (input$controlbar) {
         showModal(modalDialog(
           title = "Alert",
           "The controlbar is opened.",
           easyClose = TRUE,
           footer = NULL
         ))
       }
     })
     
     observeEvent(input$controlbarToggle, {
       updateControlbar(inputId = "controlbar")
     })
     
     observe({
       print(input$controlbar)
     })
   }
 )
}
}
