% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userListItem}
\alias{userListItem}
\title{AdminLTE2 user list item}
\usage{
userListItem(src = NULL, user_name = NULL, description = NULL)
}
\arguments{
\item{src}{image url or path.}

\item{user_name}{user name.}

\item{description}{any date element.}
}
\description{
Create a user list item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
