% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userMessage}
\alias{userMessage}
\title{AdminLTE2 user message}
\usage{
userMessage(
  ...,
  author,
  date = NULL,
  image = NULL,
  type = c("sent", "received")
)
}
\arguments{
\item{...}{Message text.}

\item{author}{Message author.}

\item{date}{Message date.}

\item{image}{Message author image path or url.}

\item{type}{Side where author is displayed: \code{c("sent", "received")}.}
}
\description{
Create a user message
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
