% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{userPostMedia}
\alias{userPostMedia}
\title{AdminLTE2 user post media}
\usage{
userPostMedia(src = NULL, height = NULL, width = NULL)
}
\arguments{
\item{src}{image path or url ...}

\item{height}{media height in pixels.}

\item{width}{media width in pixels.}
}
\description{
Create a user post media (image)
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
